refactor(git-history): Interactions 抽 PR 创建与 sync/push preview 处理器组（2878→2236）
